requirements.txt
Flask==2.3.2
httpx==0.27.0
pycryptodome==3.20.0
